monitor/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-client-cache==0.11.1
aiohttp-cors==0.7.0
aiosignal==1.4.0
aiosqlite==0.20.0
airportsdata==20241001
aistudio-sdk==0.2.6
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
arch==7.2.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
Authlib==1.6.5
babel==2.17.0
bce-python-sdk==0.9.29
beautifulsoup4==4.14.2
blake3==1.0.4
blinker==1.9.0
blobfile==3.0.0
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.1
cattrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
colorlog==6.9.0
contourpy==1.3.1
cos-python-sdk-v5==1.9.34
coscmd==1.8.6.34
courlan==1.3.2
crashtest==0.4.1
crcmod==1.7
cryptography==44.0.0
curl_adapter==1.1.0
curl_cffi==0.13.0
cycler==0.12.1
cyclopts==3.24.0
Cython==3.1.5
datasets==3.5.0
dateparser==1.2.2
DateTime==5.5
decorator==5.1.1
deepdiff==8.6.1
defusedxml==0.8.0rc2
depyf==0.18.0
dill==0.3.4
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.2
dulwich==0.22.7
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
exchange_calendars==4.11.2
executing==2.2.0
fast_dataindex==0.1.2
fastapi==0.116.2
fastjsonschema==2.21.1
fastmcp==2.12.5
filelock==3.17.0
finvizfinance==1.2.0
Flask==3.1.0
flask-babel==4.0.0
fonttools==4.57.0
formulaic==1.2.1
freezegun==1.5.5
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gevent==24.11.1
gguf==0.10.0
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.67.0
graphql-core==3.2.6
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.0.0
inscriptis==2.6.0
installer==0.7.0
interegular==0.3.3
interface-meta==1.3.0
ipython==8.32.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.8.0
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
jupyter_core==5.9.1
jusText==3.0.2
keyring==25.6.0
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
lark==1.2.2
lazy-object-proxy==1.12.0
linearmodels==6.1
llvmlite==0.45.1
lxml==5.3.2
lxml_html_clean==0.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.12.2
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.8.0
Nasdaq-Data-Link==1.0.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.17.0
openai==1.63.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openbb==4.5.0
openbb-alpha-vantage==1.5.0
openbb-benzinga==1.5.0
openbb-biztoc==1.5.0
openbb-bls==1.2.0
openbb-cboe==1.5.0
openbb-cftc==1.2.0
openbb-charting==2.4.0
openbb-commodity==1.4.0
openbb-congress-gov==1.1.0
openbb-core==1.5.2
openbb-crypto==1.5.0
openbb-currency==1.5.0
openbb-deribit==1.1.0
openbb-derivatives==1.5.0
openbb-ecb==1.5.0
openbb-econdb==1.4.0
openbb-econometrics==1.6.0
openbb-economy==1.5.0
openbb-equity==1.5.0
openbb-etf==1.5.0
openbb-famafrench==1.1.0
openbb-federal-reserve==1.5.0
openbb-finra==1.5.0
openbb-finviz==1.4.0
openbb-fixedincome==1.5.0
openbb-fmp==1.5.0
openbb-fred==1.5.0
openbb-government-us==1.5.0
openbb-imf==1.2.0
openbb-index==1.5.0
openbb-intrinio==1.5.0
openbb-mcp-server==1.2.0
openbb-multpl==1.2.0
openbb-nasdaq==1.5.0
openbb-news==1.5.0
openbb-oecd==1.5.0
openbb-platform-api==1.2.1
openbb-polygon==1.5.0
openbb-quantitative==1.5.0
openbb-regulators==1.5.0
openbb-sec==1.5.0
openbb-seeking-alpha==1.5.0
openbb-stockgrid==1.5.0
openbb-technical==1.5.0
openbb-tiingo==1.5.0
openbb-tmx==1.4.0
openbb-tradier==1.4.0
openbb-tradingeconomics==1.5.0
openbb-us-eia==1.2.0
openbb-wsj==1.5.0
openbb-yfinance==1.5.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orderly-set==5.5.0
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
paddle2onnx==2.0.0
paddlefsl==1.1.0
paddlenlp==3.0.0b4
pandas==2.2.3
pandas-ta-openbb==0.4.22
parse==1.20.2
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pathable==0.4.4
patsy==1.0.1
peewee==3.18.2
pexpect==4.9.0
pillow==11.2.1
pkginfo==1.12.0
platformdirs==4.3.6
plotly==6.3.1
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
prettytable==3.14.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==6.30.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.22.0
pycurl==7.45.7
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.1
pyhdfe==0.2.0
PyJWT==2.10.1
pyluach==2.3.0
pyparsing==3.2.3
pyperclip==1.11.0
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-freezegun==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
random-user-agent==1.0.1
RapidFuzz==3.12.1
rarfile==4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-cache==1.2.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rich-rst==1.3.2
rpds-py==0.22.3
rsa==4.9
ruff==0.13.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
SecretStorage==3.3.3
sentencepiece==0.2.0
seqeval==1.2.2
setuptools==75.8.0
setuptools-scm==8.3.1
sgqlc==16.4
shellingham==1.5.4
shinny-structlog==0.0.4
simplejson==3.19.3
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.5
sympy==1.13.1
ta==0.11.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tld==0.13.1
tokenizers==0.21.0
tomlkit==0.13.2
tool_helpers==0.1.2
toolz==1.1.0
tornado==6.4.2
tqdm==4.67.1
tqsdk==3.7.8
tqsdk-ctpse==1.0.2
tqsdk-sm==1.0.5
trafilatura==2.0.0
traitlets==5.14.3
trove-classifiers==2025.1.15.22
typer==0.15.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
url-normalize==1.4.3
urllib3==2.3.0
uuid7==0.1.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.29.1
visualdl==2.5.3
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
wheel==0.45.1
wrapt==1.17.2
xlrd==2.0.2
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.58
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
